Name the cookie a test replays, and the plan it belongs to
Forty-sixth review pass, nine findings. Eight applied, one refused.

`cookiePair` took the first `Set-Cookie` on the response. These responses can
carry more than one, so a test could replay a different plan's cookie - or a
session's - while reading as though it had proved this plan's. It now selects
by `shareCookieName(planId)` through `getSetCookie()`, and every call site says
which plan it means, including the one that deliberately uses another.

The published unlock description was contradicting the code twice over. It
still said only a refused redemption is charged, which stopped being true when
404, 400 and 413 started keeping their count; the first correction then said
"only a redemption is refunded" immediately before admitting a `500` is
refunded too. It now states the rule outright: a redemption and a `500` are
refunded, every refusal keeps its count.

The refund warning carries the bucket and window it failed to return, so a
budget that never recovers can be found rather than only known about. No
counter metric - nothing in this app consumes one, and the log line beside it
is what an operator actually reads.

README's repeat command claimed the setup block above "exports" its variables.
It does not - those are per-command assignments, so the repeat would have run
with none of them and skipped every driver it was meant to be repeating. It now
says where they have to come from and what happens without them, and it no
longer restates credentials that would drift from the block above.

Smaller: the `/s/{id}` gate records that its `noindex` comes from `Document`
omitting `social`, rather than leaving the absent `X-Robots-Tag` looking like
an oversight. `memoryStorage.put` drops a cast that was hiding nothing -
`PlanStorage.put` takes a `Uint8Array` and always did. The postgres late-client
test waits for the release instead of reading it in whichever job settled
first. The client suites derive their asserted origin from the constant
`dom-env.ts` registers the DOM with. And a code span in docs/self-hosting.md is
back on one line, which last round's rewrite had broken across two.

Refused: aliasing the stub's `SessionState` to the real one. The real type is a
`Pick` over Better Auth's own session, so every stub would have to build a
complete user record to satisfy it. Nothing in `src/` is typed against the
stub - the panels resolve `useSession` from the real module at compile time and
only the runtime is swapped - so there is no drift for the alias to catch.

+++ tests/app-harness.ts
@@ -349,7 +349,7 @@ export function memoryStorage(
       // A copy, because a real bucket keeps the bytes it was handed: retaining
       // the caller's buffer lets a later mutation of it silently rewrite a
       // stored document and the etag derived from it.
-      objects.set(id, new Uint8Array(body as Uint8Array));
+      objects.set(id, new Uint8Array(body));
     },
     get: async (id): Promise<PlanObject | null> => {
       const bytes = objects.get(id);
